[ new ] IO with bounded level polymorphism
Bounded-level polymorphism means we can use *any* `Set` under a
certain level without having to wrap them in a `Lift`.

This is relying on:

* the pragma NO_UNIVERSE_CHECK to force Agda to accept the
  definition of the codatatype `IO`.
  It is currently only available in the development version of Agda.

* instance arguments of the form `a ⊔ b ≡ a` to mean that `Level`
  `a` is smaller or equal to `Level` `b`.
  Those can be solved automatically for concrete cases and simply
  added as extra constraints on `IO`-manipulating functions.

read.agda provides an example of an IO program computing a Set from
its input (thus needing a large `IO`) but also using basic functions
such as `getLine` which are, in the standard library, forcing `IO`
to be small.